+++ Source/Scene/TexturePool.js
@@ -42,14 +42,14 @@ define([
      * <br/><br/>
      * Texture pools are useful when textures are being created and destroyed repeatedly.
      *
-     * @name Texture2DPool
+     * @name TexturePool
      * @constructor
      *
      * @param {Context} context The context to use to create textures when needed.
      *
      * @see Texture
      */
-    function Texture2DPool(context) {
+    function TexturePool(context) {
         if (typeof context === 'undefined') {
             throw new DeveloperError('context is required.');
         }
@@ -63,13 +63,13 @@ define([
      * but may return a pooled texture if there are any available.  If a pooled texture is re-used,
      * and no source is provided, the new texture will still retain its old contents.
      *
-     * @memberof Texture2DPool
+     * @memberof TexturePool
      *
      * @exception {DeveloperError} description is required.
      *
      * @see Context#createTexture2D
      */
-    Texture2DPool.prototype.createTexture2D = function(description) {
+    TexturePool.prototype.createTexture2D = function(description) {
         if (!description) {
             throw new DeveloperError('description is required.');
         }
@@ -102,13 +102,13 @@ define([
      * If this object was destroyed, it should not be used; calling any function other than
      * <code>isDestroyed</code> will result in a {@link DeveloperError} exception.
      *
-     * @memberof Texture2DPool
+     * @memberof TexturePool
      *
      * @return {Boolean} True if this object was destroyed; otherwise, false.
      *
-     * @see Texture2DPool#destroy
+     * @see TexturePool#destroy
      */
-    Texture2DPool.prototype.isDestroyed = function() {
+    TexturePool.prototype.isDestroyed = function() {
         return false;
     };
 
@@ -120,18 +120,18 @@ define([
      * <code>isDestroyed</code> will result in a {@link DeveloperError} exception.  Therefore,
      * assign the return value (<code>undefined</code>) to the object as done in the example.
      *
-     * @memberof Texture2DPool
+     * @memberof TexturePool
      *
      * @return {undefined}
      *
      * @exception {DeveloperError} This object was destroyed, i.e., destroy() was called.
      *
-     * @see Texture2DPool#isDestroyed
+     * @see TexturePool#isDestroyed
      *
      * @example
      * pool = pool && pool.destroy();
      */
-    Texture2DPool.prototype.destroy = function() {
+    TexturePool.prototype.destroy = function() {
         var free = this._free;
         Object.keys(free).forEach(function(textureTypeKey) {
             free[textureTypeKey].forEach(function(texture) {
@@ -141,5 +141,5 @@ define([
         return destroyObject(this);
     };
 
-    return Texture2DPool;
+    return TexturePool;
 });